#300: Moved classes.php to the sub-folder "includes"
 Sorting things around.
