+++ src/php/objects.php
@@ -3,7 +3,7 @@ require_once "../gestion/genscripts/object_brex_objet.class.php";
 require_once "../gestion/genscripts/object_brex_objet_categ.class.php";
 require_once "../gestion/genscripts/object_brex_obj_comp.class.php";
 require_once "../gestion/genscripts/object_brex_perso_trust.class.php";
-require_once "classes.php";
+require_once "includes/classes.php";
 require_once "skill_class.php";
 
 function dieWithBadRequest($errorMessages)
@@ -37,7 +37,7 @@ if ($_SERVER ['REQUEST_METHOD'] == 'POST') {
   $brex_perso_trust = createAndValidateObjetLienTMR($objet, $stored_brex_objet);
   storeLienTMR($brex_perso_trust);
   updateObjetWithLienTMR($stored_objet);
-  
+
   if($objet->categorie->ffchId == 67){
     saveAndRelateObjectSkills($objet, $brex_objet);
   }
@@ -167,7 +167,7 @@ function createPropertyArray($objet)
   $values ['defp'] = $objet->caracp->def;
   $values ['magp'] = $objet->caracp->mag;
   $values ['psyp'] = $objet->caracp->psy;
-  
+
   $values ['res_feu'] = $objet->elements->feu;
   $values ['res_glace'] = $objet->elements->glace;
   $values ['res_foudre'] = $objet->elements->foudre;
@@ -176,7 +176,7 @@ function createPropertyArray($objet)
   $values ['res_terre'] = $objet->elements->terre;
   $values ['res_lumiere'] = $objet->elements->lumiere;
   $values ['res_tenebres'] = $objet->elements->tenebres;
-  
+
   $values ['res_poison'] = $objet->resistancesAlterations->poison;
   $values ['res_cecite'] = $objet->resistancesAlterations->cecite;
   $values ['res_sommeil'] = $objet->resistancesAlterations->sommeil;
@@ -230,7 +230,7 @@ function updateAndValidateObjet($objet, $brex_objet)
 
 function createOrUpdateAndValidateObjet($objet, $brex_objet = null) {
   $values = createPropertyArray ( $objet );
-  
+
   if (is_null ( $brex_objet )) {
     $brex_objet = new brex_objet ( $values );
     if ($objet->categorie->ffchId == 67) {
@@ -246,12 +246,12 @@ function createOrUpdateAndValidateObjet($objet, $brex_objet = null) {
   } else {
     $brex_objet->updateObject ( $values );
   }
-  
+
   $brex_objet_categ = brex_objet_categ::findByPrimaryId ( $objet->categorie->ffchId );
   $brex_objet->setrelationcategorie ( $brex_objet_categ );
-  
+
   $brex_objet->verifyValues ();
-  
+
   return $brex_objet;
 }
 
@@ -267,28 +267,28 @@ function createAndValidateObjetLienTMR($objet, $brex_objet) {
   if (is_null ( $objet->lienTMR )) {
     return null;
   }
-  
+
   $brex_persos = brex_perso::finderParGumiId ( $objet->lienTMR->perso_gumi_id );
-  
+
   if (is_null ( $brex_persos ) || count ( $brex_persos ) != 1) {
     return null;
   }
-  
+
   $brex_perso = $brex_persos [0];
-  
+
   $values = array ();
   $values ['super'] = $objet->lienTMR->isSTMR ? '1' : '0';
   $values ['vc'] = $objet->lienTMR->isVC ? '1' : '0';
   $values ['creation_datedate'] = date ( 'Y-m-d' );
   $values ['creation_datehour'] = date ( 'H' );
   $values ['creation_datemins'] = date ( 'i' );
-  
+
   $brex_perso_trust = new brex_perso_trust ( $values );
   $brex_perso_trust->setrelationperso ( $brex_perso );
   $brex_perso_trust->setrelationreward ( $brex_objet );
-  
+
   $brex_perso_trust->verifyValues ();
-  
+
   return $brex_perso_trust;
 }
